feat: read csv file

+++ src/routes.js
@@ -6,39 +6,49 @@ const database = new Database()
 
 export const routes = [
   {
-    method: 'GET',
-    path: buildRoutePath('/tasks'),
-    handler: (req, res) => {
-      const { search } = req.query
-
-      const searchData = search ? {
-        title: search,
-        description: search
-      } : null
-
-      const tasks = database.select('tasks', searchData)
-
-      return res.end(JSON.stringify(tasks))
-    }
-  },
-  {
     method: 'POST',
     path: buildRoutePath('/tasks'),
     handler: (req, res) => {
       const { title, description } = req.body
 
-      const user = {
+      if (!title) {
+        return res.writeHead(400).end(
+          JSON.stringify({ message: 'title is required' }),
+        )
+      }
+
+      if (!description) {
+        return res.writeHead(400).end(
+          JSON.stringify({message: 'description is required' })
+        )
+      }
+
+      const task = {
         id: randomUUID(),
         title,
         description,
         completed_at: null,
         created_at: new Date(),
-        updated_at: new Date()
+        updated_at: new Date(),
       }
 
-      database.insert('tasks', user)
+      database.insert('tasks', task)
 
       return res.writeHead(201).end()
+    }
+  },
+  {
+    method: 'GET',
+    path: buildRoutePath('/tasks'),
+    handler: (req, res) => {
+      const { search } = req.query
+
+      const tasks = database.select('tasks', {
+        title: search,
+        description: search
+      })
+
+      return res.end(JSON.stringify(tasks))
     }
   },
   {
@@ -48,9 +58,22 @@ export const routes = [
       const { id } = req.params
       const { title, description } = req.body
 
+      if (!title || !description) {
+        return res.writeHead(400).end(
+          JSON.stringify({ message: 'title or description are required' })
+        )
+      }
+
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
       database.update('tasks', id, {
         title,
         description,
+        updated_at: new Date()
       })
 
       return res.writeHead(204).end()
@@ -62,6 +85,12 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
 
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
       database.delete('tasks', id)
 
       return res.writeHead(204).end()
@@ -69,11 +98,20 @@ export const routes = [
   },
   {
     method: 'PATCH',
-    path: buildRoutePath('/tasks/:id'),
+    path: buildRoutePath('/tasks/:id/complete'),
     handler: (req, res) => {
       const { id } = req.params
 
-      database.updateCompletedAt('tasks', id)
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
+      const isTaskCompleted = !!task.completed_at
+      const completed_at = isTaskCompleted ? null : new Date()
+
+      database.update('tasks', id, { completed_at })
 
       return res.writeHead(204).end()
     }
